MAINT: Move _pywt to _extensions sub-module
Creating a separate module for compiled code will be useful when splitting up
_pywt.pyx, as some functions will be moved to pure python where the names of
the modules might clash (e.g.  _dwt.py, _dwt.pyx).

This is by no means a permanent change, and more reorganization might happen.
