+++ pywt/_multilevel.py
@@ -12,8 +12,7 @@ from __future__ import division, print_function, absolute_import
 
 import numpy as np
 
-from ._pywt import Wavelet
-from ._pywt import dwt, idwt, dwt_max_level
+from ._extensions._pywt import Wavelet, dwt, idwt, dwt_max_level
 from ._multidim import dwt2, idwt2
 
 __all__ = ['wavedec', 'waverec', 'wavedec2', 'waverec2', 'iswt', 'iswt2']
