+++ pywt/_wavelet_packets.py
@@ -11,7 +11,7 @@ __all__ = ["BaseNode", "Node", "WaveletPacket", "Node2D", "WaveletPacket2D"]
 
 import numpy as np
 
-from ._pywt import Wavelet, dwt, idwt, dwt_max_level
+from ._extensions._pywt import Wavelet, dwt, idwt, dwt_max_level
 from ._multidim import dwt2, idwt2
